+++ tests/test_version.py
@@ -22,13 +22,11 @@ def test_no_extra_modules(mversion, mhash, expected):
     with contextlib.ExitStack() as stack:
         stack.enter_context(mock.patch("luxos.version.__version__", mversion))
         stack.enter_context(mock.patch("luxos.version.__hash__", mhash))
-        assert version.get_version() == {
+        assert version.get_version_info() == {
             "py": PYVERSION,
             "luxos": expected,
         }
-        assert (
-            version.get_version(as_string=True) == f"py[{PYVERSION}], luxos[{expected}]"
-        )
+        assert version.get_version() == f"py[{PYVERSION}], luxos[{expected}]"
 
 
 def test_add_module_info():
@@ -36,7 +34,7 @@ def test_add_module_info():
     with contextlib.ExitStack() as stack:
         stack.enter_context(mock.patch("luxos.version.__version__", "1.2.4"))
         stack.enter_context(mock.patch("luxos.version.__hash__", "a123b"))
-        assert version.get_version([module]) == {
+        assert version.get_version_info([module]) == {
             "py": PYVERSION,
             "luxos": "1.2.4, a123b",
             "xyz": "N/A",
@@ -46,7 +44,7 @@ def test_add_module_info():
     with contextlib.ExitStack() as stack:
         stack.enter_context(mock.patch("luxos.version.__version__", "1.2.4"))
         stack.enter_context(mock.patch("luxos.version.__hash__", "a123b"))
-        assert version.get_version([module]) == {
+        assert version.get_version_info([module]) == {
             "py": PYVERSION,
             "luxos": "1.2.4, a123b",
             "ABC": "N/A",
@@ -56,7 +54,7 @@ def test_add_module_info():
     with contextlib.ExitStack() as stack:
         stack.enter_context(mock.patch("luxos.version.__version__", "1.2.4"))
         stack.enter_context(mock.patch("luxos.version.__hash__", "a123b"))
-        assert version.get_version([module]) == {
+        assert version.get_version_info([module]) == {
             "py": PYVERSION,
             "luxos": "1.2.4, a123b",
             "ABC": "9.9.9",